refactor Doc components
