+++ web-client/src/ui/doc/components/Poor.tsx
