+++ web-client/src/ui/doc/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./DocRoot";
+export * from "./dom";
